Uses fast 8-bit int optimization in keycode -> led count func

+++ keyboards/evoworks/evo80/evo80.c
@@ -177,13 +177,13 @@ static inline bool kb_get_lock_state(void) {
 }
 
 static inline uint8_t get_led_index_for_keycode(uint16_t target_keycode, uint8_t layer) {
-    for (uint8_t row = 0; row < MATRIX_ROWS; row++) {
-        for (uint8_t col = 0; col < MATRIX_COLS; col++) {
+    for (uint_fast8_t row = 0; row < MATRIX_ROWS; row++) {
+        for (uint_fast8_t col = 0; col < MATRIX_COLS; col++) {
             uint8_t led_index = g_led_config.matrix_co[row][col];
             if (led_index == NO_LED) {
                 continue; 
             }
-            keypos_t pos = {.row = row, .col = col};
+            keypos_t pos = {.row = (uint8_t)row, .col = (uint8_t)col};
             if (keymap_key_to_keycode(layer, pos) == target_keycode) {
                 return led_index;
             }
